Rename src package to autosar_atlas_parser
