+++ setup.py
@@ -5,14 +5,14 @@ setup(
     version="0.1.0",
     package_dir={"": "src"},
     packages=find_packages("src"),
-    py_modules=["main"],
+    py_modules=[],
     install_requires=[
         # PDF処理に必要なライブラリをここに
         "pdfminer.six>=20221105"
     ],
     entry_points={
         "console_scripts": [
-            "autosar-parser=main:main",
+            "autosar-parser=autosar_atlas_parser.main:main",
         ]
     }
 )
