+++ src/autosar_atlas_parser/block_extractor.py
