+++ src/autosar_atlas_parser/section_detector.py
