+++ src/autosar_atlas_parser/trace_parser.py
